+++ assets/src/components/activities/common/stem/delivery/StemDelivery.tsx
@@ -14,7 +14,7 @@ interface StemProps {
 export const StemDelivery: React.FC<StemProps> = (props) => {
   return (
     <div className={`stem__delivery${props.className ? ' ' + props.className : ''}`}>
-      <HtmlContentModelRenderer text={props.stem.content} context={props.context} />
+      <HtmlContentModelRenderer content={props.stem.content} context={props.context} />
     </div>
   );
 };