+++ assets/src/components/activities/types.ts
@@ -29,10 +29,7 @@ export type ChoiceId = ID;
 export type PartId = ID;
 export type ResponseId = ID;
 
-export type RichText = {
-  model: ModelElement[];
-  selection: Selection;
-};
+export type RichText = ModelElement[];
 
 export interface Success {
   type: 'success';
@@ -44,16 +41,13 @@ export interface HasContent {
 export function makeContent(text: string, id?: string): { id: string; content: RichText } {
   return {
     id: id ? id : guid(),
-    content: {
-      model: [
-        create({
-          type: 'p',
-          children: [{ text }],
-          id: guid(),
-        }),
-      ],
-      selection: null,
-    },
+    content: [
+      create({
+        type: 'p',
+        children: [{ text }],
+        id: guid(),
+      }),
+    ],
   };
 }