+++ assets/src/components/editing/commands/ImageCmd.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import { Transforms } from 'slate';
 import { ReactEditor } from 'slate-react';
-import * as ContentModel from 'data/content/model';
 import { MIMETYPE_FILTERS, SELECTION_TYPES } from 'components/media/manager/MediaManager';
 import ModalSelection, { sizes } from 'components/modal/ModalSelection';
 import { modalActions } from 'actions/modal';
 import { MediaItem } from 'types/media';
 import { Command, CommandDesc } from 'components/editing/commands/interfaces';
 import { UrlOrUpload } from 'components/media/UrlOrUpload';
-import { Maybe } from 'tsmonad';
 import guid from 'utils/guid';
 
 const dismiss = () => (window as any).oliDispatch(modalActions.dismiss());
@@ -16,9 +14,9 @@ const display = (c: any) => (window as any).oliDispatch(modalActions.display(c))
 
 export function selectImage(
   projectSlug: string,
-  selectedUrl?: string,
+  _selectedUrl?: string,
 ): Promise<string | undefined> {
-  return new Promise((resolve, reject) => {
+  return new Promise((resolve, _reject) => {
     let selectedUrl: string | undefined = undefined;
 
     const mediaLibrary = (
@@ -37,7 +35,6 @@ export function selectImage(
           onUrlChange={(url: string) => (selectedUrl = url)}
           onMediaSelectionChange={(mediaOrUrl: MediaItem[]) => (selectedUrl = mediaOrUrl[0]?.url)}
           projectSlug={projectSlug}
-          // eslint-disable-next-line
           onEdit={() => {}}
           mimeFilter={MIMETYPE_FILTERS.IMAGE}
           selectionType={SELECTION_TYPES.SINGLE}
@@ -50,30 +47,18 @@ export function selectImage(
   });
 }
 
-const command: Command = {
-  execute: (context, editor) => {
-    const at = editor.selection as any;
-    selectImage(context.projectSlug).then((img) =>
-      Maybe.maybe(img).caseOf({
-        just: (img: string) => Transforms.insertNodes(editor, ContentModel.image(img), { at }),
-        // eslint-disable-next-line
-        nothing: () => {},
-      }),
-    );
-  },
-  precondition: (editor) => {
-    return true;
-  },
-};
-
 const libraryCommand: Command = {
   execute: (context, editor: ReactEditor) => {
-    const at = editor.selection as any;
+    const at = editor.selection;
     selectImage(context.projectSlug, undefined).then((src) =>
-      Transforms.insertNodes(editor, { type: 'img', src, guid: guid(), children: [] }, { at }),
+      Transforms.insertNodes(
+        editor,
+        { type: 'img', src, guid: guid(), children: [] },
+        at ? { at } : undefined,
+      ),
     );
   },
-  precondition: (editor: ReactEditor) => {
+  precondition: (_editor) => {
     return true;
   },
 };
@@ -81,7 +66,7 @@ const libraryCommand: Command = {
 function createCustomEventCommand(onRequestMedia: (r: any) => Promise<string | boolean>) {
   const customEventCommand: Command = {
     execute: (context, editor: ReactEditor) => {
-      const at = editor.selection as any;
+      const at = editor.selection;
 
       const request = {
         type: 'MediaItemRequest',
@@ -96,11 +81,11 @@ function createCustomEventCommand(onRequestMedia: (r: any) => Promise<string | b
             id: guid(),
             children: [],
           };
-          Transforms.insertNodes(editor, img, { at });
+          Transforms.insertNodes(editor, img, at ? { at } : undefined);
         }
       });
     },
-    precondition: (editor: ReactEditor) => {
+    precondition: (_editor) => {
       return true;
     },
   };