+++ assets/src/components/editing/models/image/Editor.tsx
@@ -46,7 +46,6 @@ export const ImageEditor = (props: ImageProps): JSX.Element => {
         <HoveringToolbar
           isOpen={() => focused && selected}
           showArrow
-          // parentRef={parentRef}
           target={
             <div>
               {ReactEditor.isFocused(editor) && selected && imageRef.current && (